Split DragAndScale
